backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["riisatoro <riisa.toro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
pymongo = "^4.1.1"
PyJWT = "^2.3.0"
email-validator = "^1.2.1"
pytz = "^2022.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
ipython = "^8.3.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
